[Decomp][Player] Cohesion split: extract item-apply system to PlayerItemApply.cpp
